[modelviews] remove restriction on basemodel view subclassing.
