+++ dbflow/src/main/java/com/raizlabs/android/dbflow/structure/BaseModelView.java
@@ -2,7 +2,6 @@ package com.raizlabs.android.dbflow.structure;
 
 import com.raizlabs.android.dbflow.annotation.ModelView;
 import com.raizlabs.android.dbflow.annotation.ModelViewQuery;
-import com.raizlabs.android.dbflow.config.FlowManager;
 
 /**
  * Description: Provides a base implementation for a ModelView. Use a {@link ModelView}
@@ -10,16 +9,5 @@ import com.raizlabs.android.dbflow.config.FlowManager;
  * field via {@link ModelViewQuery}.
  */
 public abstract class BaseModelView extends NoModificationModel {
-
-    private transient RetrievalAdapter adapter;
-
-    @SuppressWarnings("unchecked")
-    @Override
-    public RetrievalAdapter getRetrievalAdapter() {
-        if (adapter == null) {
-            adapter = FlowManager.getModelViewAdapter(getClass());
-        }
-        return adapter;
-    }
 
 }
