+++ dbflow/src/main/java/com/raizlabs/android/dbflow/structure/ModelViewAdapter.java
@@ -7,7 +7,7 @@ import com.raizlabs.android.dbflow.config.DatabaseDefinition;
 /**
  * Description: The base class for a {@link TModelView} adapter that defines how it interacts with the DB.
  */
-public abstract class ModelViewAdapter<TModelView extends BaseModelView>
+public abstract class ModelViewAdapter<TModelView>
         extends InstanceAdapter<TModelView> {
 
     public ModelViewAdapter(DatabaseDefinition databaseDefinition) {
